+++ src/layers/regressionlayer.js
@@ -40,7 +40,7 @@ goog.scope(function() {
   pro.backward = function(y) {
     // compute and accumulate gradient wrt weights and bias of this layer
     var x = this.in_act;
-    x.dw = new Float32Array(x.w.length); // zero out the gradient of input Vol
+    x.dw = new Float64Array(x.w.length); // zero out the gradient of input Vol
     var loss = 0.0;
     if(y instanceof Array || y instanceof Float64Array) {
       for(var i=0;i<this.out_depth;i++) {
