+++ src/layers/relulayer.js
@@ -46,7 +46,7 @@ goog.scope(function() {
     var V = this.in_act; // we need to set dw of this
     var V2 = this.out_act;
     var N = V.w.length;
-    V.dw = new Float32Array(N); // zero out gradient wrt data
+    V.dw = new Float64Array(N); // zero out gradient wrt data
     for(var i=0;i<N;i++) {
       if(V2.w[i] <= 0) V.dw[i] = 0; // threshold
       else V.dw[i] = V2.dw[i];
