+++ src/layers/svmlayer.js
@@ -37,7 +37,7 @@ goog.scope(function() {
   pro.backward = function(y) {
     // compute and accumulate gradient wrt weights and bias of this layer
     var x = this.in_act;
-    x.dw = new Float32Array(x.w.length); // zero out the gradient of input Vol
+    x.dw = new Float64Array(x.w.length); // zero out the gradient of input Vol
 
     // we're using structured loss here, which means that the score
     // of the ground truth should be higher than the score of any other
